+++ libs/docs/core/time-picker/examples/time-picker-format-example.component.ts
@@ -1,7 +1,14 @@
 import { DatePipe } from '@angular/common';
 import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { FdDate, provideDateTimeFormats } from '@fundamental-ngx/core/datetime';
+import { DatetimePickerComponent } from '@fundamental-ngx/core';
+import {
+    DateTimeFormatPipe,
+    DayPeriodFormatPipe,
+    FdDate,
+    provideDateTimeFormats,
+    TranslateDayPeriodPipe
+} from '@fundamental-ngx/core/datetime';
 import { TimePickerModule } from '@fundamental-ngx/core/time-picker';
 
 @Component({
@@ -9,10 +16,22 @@ import { TimePickerModule } from '@fundamental-ngx/core/time-picker';
     templateUrl: './time-picker-format-example.component.html',
     providers: [provideDateTimeFormats()],
     standalone: true,
-    imports: [TimePickerModule, FormsModule, DatePipe]
+    imports: [
+        TimePickerModule,
+        FormsModule,
+        DatePipe,
+        DayPeriodFormatPipe,
+        DateTimeFormatPipe,
+        DatetimePickerComponent,
+        TranslateDayPeriodPipe
+    ]
 })
 export class TimePickerFormatExampleComponent {
-    time = new FdDate().setTime(12, 0, 0);
-    // FdDatetimeAdapter is based on Intl.DateTimeFormat.
-    displayFormat = { hour: 'numeric', minute: '2-digit', second: '2-digit', hourCycle: 'h23' };
+    time = FdDate.getNow();
+    // FdDatetimeAdapter is based Intl.DateTimeFormatOptions
+    basicDateTimeFormat = { hour: 'numeric', minute: '2-digit', second: '2-digit', hourCycle: 'h23' };
+    dateTimeFormatWithDayPeriod = {
+        ...this.basicDateTimeFormat,
+        dayPeriod: true
+    };
 }